Restructure; adding more Paths
