+++ GameEngine/src/main/java/org/euph/engine/util/Path.java
@@ -3,5 +3,8 @@ package org.euph.engine.util;
 public class Path {
     public static final String ROOT = "/";
     public static final String RESOURCES = ROOT + "res/";
-    public static  final String SHADERS = RESOURCES + "shaders/";
+    public static final String SHADERS = RESOURCES + "shaders/";
+    //Default Paths
+    public static final String DEFAULT_RESOURCES = RESOURCES + "default/";
+    public static final String DEFAULT_SHADERS = DEFAULT_RESOURCES + "shaders/";
 }
